# dashboard/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "larkstack console API",
    "description": "Admin API for the larkstack host: subsystem status, live config, an event stream (SSE), the Lark-app registry, action dispatch, and Lark OAuth login.",
    "contact": {
      "name": "S3anJia"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/actions/{app}/{action}": {
      "post": {
        "tags": [
          "console"
        ],
        "summary": "`POST /api/actions/{app}/{action}` — fire-and-forget an App action. The\nresult string from `Instance::handle_action` surfaces on the event stream,\nnot in this response.",
        "operationId": "dispatch_action",
        "parameters": [
          {
            "name": "app",
            "in": "path",
            "description": "Registered app name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "in": "path",
            "description": "Action name from the app's manifest",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Optional action parameters (JSON object)",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Action dispatched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionAck"
                }
              }
            }
          },
          "404": {
            "description": "Unknown app",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "App is not running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/events": {
      "get": {
        "tags": [
          "console"
        ],
        "summary": "`GET /api/events` — backfill recent events, then stream live ones. Honors\n`?since=<id>` and the `Last-Event-ID` header for reconnect backfill.",
        "description": "Each message's `data` is a JSON `Event`\n`{ id, level, subsystem, target, message, fields, timestamp }`. A `lagged`\nevent signals the client fell behind and some events were dropped.",
        "operationId": "stream_events",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "Replay events with id greater than this (same role as `Last-Event-ID`).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of console events; each message's `data` is one JSON `Event`",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionAck": {
        "type": "object",
        "description": "Acknowledgement that an action was queued for the running App.",
        "required": [
          "ok",
          "app",
          "action"
        ],
        "properties": {
          "action": {
            "type": "string"
          },
          "app": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Error body shared by every non-2xx JSON response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Event": {
        "type": "object",
        "description": "Wire shape of one SSE `data` payload — mirrors [`larkstack_core::Event`]'s\nserde output, declared here so the schema lands in the OpenAPI spec without\n`larkstack-core` depending on `utoipa`. Never constructed: the handler\nserializes the core event directly.",
        "required": [
          "id",
          "level",
          "target",
          "message",
          "fields",
          "timestamp"
        ],
        "properties": {
          "fields": {
            "type": "object",
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            }
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "level": {
            "$ref": "#/components/schemas/Level"
          },
          "message": {
            "type": "string"
          },
          "subsystem": {
            "type": [
              "string",
              "null"
            ]
          },
          "target": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "Unix milliseconds.",
            "minimum": 0
          }
        }
      },
...
      "Level": {
        "type": "string",
        "description": "Mirrors [`larkstack_core::Level`]'s lowercase serde encoding.",
        "enum": [
          "trace",
          "debug",
          "info",
          "warn",
          "error"
        ]
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lk_session"
      }
    }
  },
...
}